customer-facing-system/intent-recognition-engine/actions/requirements.txt
greenlet==1.1.2
python-dotenv==0.20.0
SQLAlchemy==1.4.36
